% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdf.tnb.ds.R
\name{cdf.tnb.ds}
\alias{cdf.tnb.ds}
\title{A function to calculate cdf of Truncated Negative Binomial}
\usage{
cdf.tnb.ds(y, mu, size, lower.tail = FALSE, log.p = FALSE)
}
\arguments{
\item{y}{y values.}

\item{mu}{mu parameter of TNB distribution.}

\item{size}{size parameter of TNB distribution.}
}
\description{
A function to calculate cdf of Truncated Negative Binomial
}
