% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdf.tp.li.R
\name{cdf.tp.li}
\alias{cdf.tp.li}
\title{A function to calculate cdf of Truncated Poisson}
\usage{
cdf.tp.li(y, lambda, lower.tail = FALSE, log.p = FALSE)
}
\arguments{
\item{y}{y values.}

\item{lambda}{lambda parameter of TP distribution.}
}
\description{
A function to calculate cdf of Truncated Poisson
}
