% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dhurdle.nb.ds.R
\name{dhurdle.nb.ds}
\alias{dhurdle.nb.ds}
\title{A function to calculate pdf of hurdle negative binomial.}
\usage{
dhurdle.nb.ds(y, mu, size, pi, log = FALSE)
}
\arguments{
\item{y}{y values.}

\item{mu}{mu parameter.}

\item{size}{parameter.}

\item{pi}{hu parameter.}
}
\description{
A function to calculate pdf of hurdle negative binomial.
}
