% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dhurdle.poisson.ds.R
\name{dhurdle.poisson.ds}
\alias{dhurdle.poisson.ds}
\title{A function to calculate pdf of hurdle poisson.}
\usage{
dhurdle.poisson.ds(y, lambda, pi, log = FALSE)
}
\arguments{
\item{y}{y values.}

\item{lambda}{lambda parameter.}

\item{pi}{hu parameter.}
}
\description{
A function to calculate pdf of hurdle poisson.
}
