% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.pred.dist.HP.R
\name{log.pred.dist.HP}
\alias{log.pred.dist.HP}
\title{A function to calculate posterior predictive log hurdle poisson pmf and cdf of 'brm' fit.}
\usage{
\method{log}{pred.dist.HP}(fit)
}
\arguments{
\item{fit}{A \code{brm} fit.}
}
\description{
A function to calculate posterior predictive log hurdle poisson pmf and cdf of 'brm' fit.
}
