% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmes.data.R
\docType{data}
\name{nmes.data}
\alias{nmes.data}
\title{Example dataset: nmes.data}
\format{
A data frame with 4406 observations and 22 variables
}
\source{
https://journaldata.zbw.eu/dataset/demand-for-medical-care-by-the-elderly-a-finite-mixture-approach
}
\usage{
nmes.data
}
\description{
A example dataset with 22 variables:
\enumerate{
\item ofp
\item ofnp
\item opp
\item opnp
\item emr
\item hosp
\item exclhlth
\item poorhlth
\item numchron
\item adldiff
\item noreast
\item midwest
\item west
\item age
\item black
\item male
\item married
\item school
\item faminc
\item employed
\item privins
\item medicaid
}
}
\keyword{datasets}
