% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf.tnb.ds.R
\name{pdf.tnb.ds}
\alias{pdf.tnb.ds}
\title{A function to calculate pdf of Truncated Negative Binomial}
\usage{
pdf.tnb.ds(y, mu, size, log.p = FALSE)
}
\arguments{
\item{y}{y values.}

\item{mu}{mu parameter of TNB distribution.}

\item{size}{size parameter of TNB distribution.}
}
\description{
A function to calculate pdf of Truncated Negative Binomial
}
