% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf.tp.li.R
\name{pdf.tp.li}
\alias{pdf.tp.li}
\title{A function to calculate pdf of Truncated Poisson}
\usage{
pdf.tp.li(y, lambda, log.p = FALSE)
}
\arguments{
\item{y}{y values.}

\item{lambda}{lambda parameter of TP distribution.}
}
\description{
A function to calculate pdf of Truncated Poisson
}
