% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phurdle.pois.li.R
\name{phurdle.pois.li}
\alias{phurdle.pois.li}
\title{A function to calculate predictive probability of hurdle poisson.}
\usage{
phurdle.pois.li(y, lambda, pi, lower.tail = FALSE, log.p = FALSE)
}
\arguments{
\item{y}{y values.}

\item{lambda}{lambda parameter.}

\item{pi}{hu parameter.}
}
\description{
A function to calculate predictive probability of hurdle poisson.
}
