% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zresidual_hurdle_poisson.R
\name{zresidual_hurdle_poisson}
\alias{zresidual_hurdle_poisson}
\title{A function to calculate z-residuals of a 'brm' fit.}
\usage{
zresidual_hurdle_poisson(fit, type, method = "iscv", nrep = 1)
}
\arguments{
\item{fit}{A \code{brm} fit.}

\item{type}{The component which the z-residuals should be calculated.}
}
\description{
A function to calculate z-residuals of a 'brm' fit.
}
