% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zresidual_poisson.R
\name{zresidual_poisson}
\alias{zresidual_poisson}
\title{A function to calculate z-residuals of a 'brm' fit.
This function is to be used when the user needs to calculate the z-residuals of TNB/HNB}
\usage{
zresidual_poisson(fit, type = "pois", method = "iscv", nrep = 1)
}
\arguments{
\item{fit}{A \code{brm} fit.}

\item{type}{The component which the z-residuals should be calculated.}
}
\description{
A function to calculate z-residuals of a 'brm' fit.
This function is to be used when the user needs to calculate the z-residuals of TNB/HNB
}
